fix: Resolve MultiQC callback errors by separating view and design mode callbacks
Fixes callback errors when viewing/editing dashboards with MultiQC components.

**Problem:**
- Callbacks expecting selector components (multiqc-module-select, etc.) were
  registered in both viewer and editor apps
- Saved MultiQC components don't have selectors (only created in stepper)
- Dash tried to initialize callbacks for saved components → error

**Solution:**
1. Moved three design-mode callbacks from core.py to design.py:
   - load_multiqc_metadata
   - populate_plot_and_dataset_selectors
   - render_multiqc_plot

2. Set prevent_initial_call=True on all design callbacks to prevent
   initialization for saved components

3. Cleaned up unused imports in core.py

**Result:**
- Viewer app: Only registers view-mode callbacks (no selector dependencies)
- Editor app: Registers both view-mode + design-mode callbacks
- Saved components render without errors
- Stepper components work correctly with selectors
